requirements.txt
langchain_openai 
langchain_core
python-dotenv
langchain
streamlit
langchain_community
langchain-groq
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
arxiv
wikipedia
langchainhub
text_generation
transformers
langgraph
google-search-results